feat: Change the history file format
The history file was previously a tab separated file with two columns.
The first column indicated the flash card and the second indicated
whether the user answered correctly.

This change adds a column that indicates the line type to make it easier
to add other status messages in the future.